Remove whole build block from docker compose before convox deploy
Previously this was being done with `sed`, but this would break with an
unusual docker compose file, especially the `args` key since that
contains an array of values which would be very difficult to do with
`sed`. Instead, we are now actually parsing the yaml using ruby and
removing the entire key before writing it back to the file.